More work on machines
