+++ src/main/java/gregicadditions/recipes/categories/machines/MachineRecipeOverride.java
@@ -1,19 +1,13 @@
 package gregicadditions.recipes.categories.machines;
 
 import gregicadditions.GAConfig;
-import gregicadditions.GAValues;
-import gregicadditions.item.GAMachineCasing;
-import gregicadditions.item.GAMetaBlocks;
 import gregicadditions.item.GAMetaItems;
 import gregicadditions.machines.GATileEntities;
 import gregtech.api.GTValues;
 import gregtech.api.items.OreDictNames;
 import gregtech.api.recipes.ModHandler;
-import gregtech.api.recipes.RecipeMaps;
 import gregtech.api.unification.OreDictUnifier;
 import gregtech.api.unification.material.MarkerMaterials;
-import gregtech.api.unification.material.Materials;
-import gregtech.api.unification.ore.OrePrefix;
 import gregtech.api.unification.stack.UnificationEntry;
 import gregtech.common.ConfigHolder;
 import gregtech.common.blocks.BlockMachineCasing;
@@ -30,20 +24,12 @@ import java.util.List;
 
 import static gregicadditions.GAEnums.GAOrePrefix.gtMetalCasing;
 import static gregicadditions.GAMaterials.*;
-import static gregicadditions.GAMaterials.Zylon;
-import static gregicadditions.GAValues.EV;
-import static gregicadditions.GAValues.HV;
-import static gregicadditions.GAValues.IV;
-import static gregicadditions.GAValues.LuV;
-import static gregicadditions.GAValues.UV;
-import static gregicadditions.GAValues.ZPM;
 import static gregicadditions.recipes.helper.AdditionMethods.*;
 import static gregicadditions.recipes.helper.GACraftingComponents.*;
 import static gregtech.api.GTValues.*;
 import static gregtech.api.recipes.RecipeMaps.ASSEMBLER_RECIPES;
 import static gregtech.api.recipes.ingredients.IntCircuitIngredient.getIntegratedCircuit;
 import static gregtech.api.unification.material.MarkerMaterials.Tier.*;
-import static gregtech.api.unification.material.MarkerMaterials.Tier.Master;
 import static gregtech.api.unification.material.Materials.*;
 import static gregtech.api.unification.ore.OrePrefix.*;
 import static gregtech.common.blocks.BlockMetalCasing.MetalCasingType.BRONZE_BRICKS;
